Use namespace as instance label
If we set only the job name and not the instance label, then multiple
copies of this running in different namespaces will override each other
-- only the job name is used as the group in pushgateway, and when
something in another namespace pushes metrics with the same group, they
override the ones already there.

However, there is the notion of an "instance" label which is also used
as part of the group, so by using the JOB_NAMESPACE as the instance
label we can keep multiple copies in different namespaces from
overriding each other.