Bugfix: Send instance label with URL
For instance to be used as part of the group, it needs to be included
with the URL we post to. While other labels can be used as part of the
group, it appears customary to use instance. If we need additional
groups, we could add support for arbitrary group labels here.

* Stricter tests around where we expect instance to appear
* New test ensuring instances don't clobber each other